Ready for demo member list sorted Borrow cancel clears page member add member clears fields member balance updated correctly Added settings table and actual settings

+++ toybox/views/borrow.py
@@ -30,16 +30,13 @@ def borrow(request, member_id):
     # base page context
     context.update({"daily_balance": 23.20, "login_name": "Jess Benning"})
 
-    # display toy in toy summary
+     # display toy in toy summary
     context.update(handle_toy_summary(request))
 
     # handle member search
     context.update(handle_member_search(request))
 
-    # if member_id set display member summary and list of borrowed toys
-    if (member_id):
-        context.update(handle_member_summary(request, member_id))
-        context.update(handle_borrowed_toy_list(request, member_id))
+
 
     # Always need this so search box renders
     context.update(handle_toy_borrow(request, member_id))
@@ -48,7 +45,13 @@ def borrow(request, member_id):
     context.update({"new_borrow_list": new_borrow_list})
 
     context.update(handle_payment_form(request, member_id))
-    print(context)
+    #print(context)
+
+     # if member_id set display member summary and list of borrowed toys
+    if (member_id):
+        context.update(handle_member_summary(request, member_id))
+        context.update(handle_borrowed_toy_list(request, member_id))
+
     return render(request, 'toybox/borrow.html', context)
 
 
@@ -61,7 +64,7 @@ def handle_toy_borrow(request, member_id):
     form = ToySearchForm(request.POST)
 
     if (request.method == "POST"):
-        print("POST")
+        #print("POST")
         if form.is_valid():
             if "add_toy" in request.POST:
                 toycode = form.cleaned_data['toy_id']
@@ -113,7 +116,7 @@ def handle_payment_form(request, member_id):
     # TODO update member balance and transaction table
     # TODO update toy loan duration
     context = {}
-# TODO store default duration in settings. Request.POST here ruins initial data
+
     #payment_form = PaymentForm(request.POST)#initial={'loan_duration':'2'})#request.POST)
 
     if (request.method == "POST"):
@@ -130,43 +133,52 @@ def handle_payment_form(request, member_id):
                 TempBorrowList.objects.filter(toy__code=toy_to_remove, member__id=member_id).delete()
 
             if item == "done":
-
-
                 if payment_form.is_valid():
-                    # print("valid form")
+                    print("valid form")
                     member = get_object_or_404(Member, pk=member_id)
 
-                    for new_toy in new_borrow_list:
-                        loan_duration = payment_form.cleaned_data['loan_duration']
-                        # print("LOAN_DURATION: "+loan_duration)
-                        toy = get_object_or_404(Toy, code=new_toy.toy.code)
-                        toy.borrow_toy(member, int(loan_duration))
-                        TempBorrowList.objects.filter(toy__code=new_toy.toy.code, member__id=member_id).delete()
+                    if new_borrow_list.count()>0:
+                        for new_toy in new_borrow_list:
+                            default_loan_duration = payment_form.cleaned_data['loan_duration']
+                            # print("LOAN_DURATION: "+loan_duration)
+                            toy = get_object_or_404(Toy, code=new_toy.toy.code)
+                            toy.borrow_toy(member, int(default_loan_duration))
+                            TempBorrowList.objects.filter(toy__code=new_toy.toy.code, member__id=member_id).delete()
 
-                    try:
-                        fee_due = decimal.Decimal(payment_form.cleaned_data['fee_due'])
-                    except:
-                        due_error = "Not a number"
-                        # print(payment_form.cleaned_data['fee_due']," ",due_error)
-                        break
+                        try:
+                            fee_due = decimal.Decimal(payment_form.cleaned_data['fee_due'])
+                        except:
+                            due_error = "Not a number"
+                            # print(payment_form.cleaned_data['fee_due']," ",due_error)
+                            break
 
-                    print("Due " + str(fee_due))
-                    # TODO add error message
-                    try:
-                        fee_paid = decimal.Decimal(payment_form.cleaned_data['fee_paid'])
-                    except:
-                        paid_error = "Not a number"
-                        # print(payment_form.cleaned_data['fee_paid']," ",paid_error)
-                        break
+                        print("Due " + str(fee_due))
+                        # TODO add error message
+                        try:
+                            fee_paid = decimal.Decimal(payment_form.cleaned_data['fee_paid'])
+                        except:
+                            paid_error = "Not a number"
+                            # print(payment_form.cleaned_data['fee_paid']," ",paid_error)
+                            break
 
-                    # print("Paid "+str(fee_paid))
-                    # TODO create transaction record
-                    # print(member.balance)
-                    member.balance = member.balance - fee_due + fee_paid
-                    print(member.balance)
-                    member.save()
+                        # print("Paid "+str(fee_paid))
+                        # TODO create transaction record
+                        # print(member.balance)
+                        member.balance = member.balance - fee_due + fee_paid
+                        #print(member.balance)
+                        member.save()
+                else:
+                    print("invalid form")
 
-    payment_form = PaymentForm(initial={'loan_duration':'2'})
+    try:
+        default_loan_duration = Settings.objects.get(name="default_loan_duration").value
+    except Settings.DoesNotExist:
+        default_loan_duration = "2"
+
+    # TODO Request.POST here ruins initial data
+    #payment_form = PaymentForm(request.POST,initial={'loan_duration':'2'})
+    #including request.POST clears initial data, without it validation errors are missing
+    payment_form = PaymentForm(initial={'loan_duration':default_loan_duration})
 
     context.update({'payment_form': payment_form})
 
@@ -176,15 +188,18 @@ def handle_payment_form(request, member_id):
 
 class PaymentForm(forms.Form):
     numeric = RegexValidator(r'^[0-9.]*$', 'Only numeric characters are allowed.')
-    # TODO store loan durations options in settings
-    LOAN_DURATIONS = "126"
-    LOAN_CHOICES = []
 
+    try:
+        loan_durations = Settings.objects.get(name="loan_durations").value
+    except Settings.DoesNotExist:
+        loan_durations = "126"
+
+    loan_choices = []
     # so choices can be easily stored in settings
-    for c in LOAN_DURATIONS:
-        LOAN_CHOICES.append((c, c))
+    for c in loan_durations:
+        loan_choices.append((c, c))
 
-    loan_duration = forms.ChoiceField(choices=LOAN_CHOICES, widget=forms.RadioSelect())
+    loan_duration = forms.ChoiceField(choices=loan_choices, widget=forms.RadioSelect())
     fee_due = forms.CharField(label="Fee Due", max_length=20, validators=[numeric])
     fee_paid = forms.CharField(label="Fee Paid", max_length=20, validators=[numeric])
